Tried to get a lunar intercept, found a crash instead
The problem is that I was using min instead of max, and so sometimes
there really wouldn't be a root. But if I do max, then there are cases
where a ship would be right on the edge of an SOI, and we'd get stuck in
a cycle of "do SOI encounter, do SOI exit", and stack overflow.

So instead we just don't generate an event if we're already in the SOI.

Also, some touchups:
- Add num_iterations to find_root_bracket
- Actually use sibling_bodies (oops!)
- Add another quick check to SOI encounter (relative velocity)