Fix orbit patch rendering problems
This really isn't ideal. The end_anomaly field tells us what s will be
when the event happens, but we can only find s modulo the orbit period,
so we have to do shenanigans to adjust.

I'm not sure this is much good either -- if the event occurs more than
one period from now I think it'll render wrong. More thinking is
required.