Create `Timeline` object
I'm having some trouble with rewinding and replaying not being a no-op.
Starting to think that simulating by timestep is not the right play.

Instead, I'm going to build up a timeline of the universe as a sequence
of segments, where within each segment, no events happen, and everything
moves in nice, Keplerian orbits.

Then the simulation just checks the timeline for what it should render.